requirements.txt
flask>=3.1.0
opencv-python>=4.13
tflite-runtime>=2.14
numpy>=1.23
pyserial>=3.6
RPi.GPIO>=0.7.1
